+++ src/acp/util/getAgentAccessAll.ts
@@ -45,7 +45,7 @@ import type { WithAccessibleAcr } from "../acp";
  *
  * @param resourceUrl URL of the Resource you want to read the access for.
  * @param options Default Options such as a fetch function.
- * @since unreleased
+ * @since 1.21.0
  */
 export async function getAgentAccessAll<T extends WithAccessibleAcr>(
   resourceWithAcr: T