+++ CHANGELOG.md
@@ -11,6 +11,11 @@ The following changes have been implemented but not released yet:
 - `getAgentAccessAll` has been added to the new `universal` access module. This
   function provides an overview of access modes granted to all agents.
 
+### Bugfixes
+
+- `getProfileAll` and `getPodUrlAll` no longer make an authenticated request to the
+  WebID profile, which should be a public resource in the first place.
+
 ## [1.20.2] - 2022-03-18
 
 ### Bugfixes